Make CUDA implementation C++11 compatible (#1)
* Make CUDA implementation C++11 compatible

By splitting the definition of nested namespaces the CUDA implementation becomes compatible with C++11. Besides this, the current implementation does not use any C++>11 features. Therefore, this change allows testing the code on machines with older NVCC versions.

* Add dev branch to CI pipeline

Format checks are now run during CI for the dev branch.

* Add instructions on how to configure the project

Before one can run CMake, the git submodules have to be downloaded and the CMakeUserPresets.json has to be defined.