Add Baseline and first paper draft (#5)
This pull request adds a first draft of the paper. In the paper, we compare the performance of the different variants with a baseline that we added to the source code as well. Furthermore, we set up a CI pipeline to render the notebooks and show them on doc pages.